move code to inser pause and change recipient into dedicated class
goal: have roughtly the same level of abrastrion among the different
method of SubscriptionCommand

also clean up dangling comments etc

Pause subscription delivery for given timeframe #2495
