+++ de.metas.contracts/src/main/java/de/metas/flatrate/api/impl/ContractChangeBL.java
@@ -72,10 +72,11 @@ public class ContractChangeBL implements IContractChangeBL
 
 		final IOrderPA orderPA = Services.get(IOrderPA.class);
 
-		final ISubscriptionDAO subscriptionPA = Services.get(ISubscriptionDAO.class);
+		final ISubscriptionDAO subscriptionDAO = Services.get(ISubscriptionDAO.class);
 
 		// get the subscription progress entries that we might be dealing with
-		final List<I_C_SubscriptionProgress> sps = subscriptionPA.retrieveSubscriptionProgresses(SubscriptionProgressQuery.builder().term(currentTerm).build());
+		final SubscriptionProgressQuery currentTermQuery = SubscriptionProgressQuery.term(currentTerm).build();
+		final List<I_C_SubscriptionProgress> sps = subscriptionDAO.retrieveSubscriptionProgresses(currentTermQuery);
 
 		final Properties ctx = InterfaceWrapperHelper.getCtx(currentTerm);
 		final String trxName = InterfaceWrapperHelper.getTrxName(currentTerm);
